Fix Build 💚 - Rename StructureMapUtilities unit tests
